Put clash test case in its own file
